# packages/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResiliencePay API",
    "description": "API for the AI Revenue Recovery Agent \u2014 Detect, Diagnose, Decide, Act, Observe",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/events/ingest": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Ingest Event",
        "description": "Ingests a new failure event and queues for diagnosis.",
        "operationId": "ingest_event_v1_events_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "description": "Returns the full pipeline state for one event using DTO mapper.",
        "operationId": "get_event_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStateDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events/{event_id}/mark-resolved": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Mark Resolved",
        "description": "Manual trigger for live demos only.",
        "operationId": "mark_resolved_v1_events__event_id__mark_resolved_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkResolvedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EventStateDTO": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "cause_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause Category"
          },
          "diagnosis_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis Confidence"
          },
          "chosen_arm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Arm"
          },
          "gate_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Passed"
          },
          "simulated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulated"
          },
          "outcome_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Result"
          },
          "amount_recovered": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Recovered"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "episode_id",
          "cause_category",
          "diagnosis_confidence",
          "chosen_arm",
          "gate_passed",
          "simulated",
          "outcome_result",
          "amount_recovered",
          "occurred_at"
        ],
        "title": "EventStateDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestEventRequest": {
        "properties": {
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "default": "subscription_charge_failed"
          },
          "merchant_id": {
            "type": "string",
            "title": "Merchant Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "INR"
          },
          "gateway_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Error Code"
          },
          "raw_gateway_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Gateway Message"
          },
          "customer_segment": {
            "type": "string",
            "title": "Customer Segment",
            "default": "new"
          },
          "retry_count_so_far": {
            "type": "integer",
            "title": "Retry Count So Far",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "merchant_id",
          "customer_id",
          "amount"
        ],
        "title": "IngestEventRequest"
      },
      "MarkResolvedRequest": {
        "properties": {
          "result": {
            "type": "string",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "MarkResolvedRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}